fix warnings.
git-svn-id: svn://qnap.local/TwoTerm/branches/fix-gno-scrolling-region@3158 5590a31f-7b70-45f8-8c82-aa3a8e5f4507

+++ Views/EmulatorView.mm
@@ -570,7 +570,7 @@
     std::string rv;
     int offset = 0;
     while (offset < bytes.size()) {
-        int max = bytes.size() - offset;
+        size_t max = bytes.size() - offset;
         if (max > 16) max = 16;
         
         rv.append(hex.data() + offset * 3, max * 3);
